backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.5
pytesseract==0.3.10
pdfplumber==0.9.0
Pillow==10.1.0
python-docx==1.1.0
reportlab==4.0.7
requests==2.31.0